Implement helper to adjust RAM-related configs
Sharing the same config between beefy desktop and thin and weak laptop
results in a suboptimal experience (either underpowered on desktop
or over-taxed on laptop).

Add an example of how to run the function at config-eval time and use
its result in a setq sexp. Still possible to reset the config with
dir-locals for specific projects, but sets a more reasonable default